requirements.txt
pandas
openpyxl==3.1.2
requests==2.31.0
schedule==1.2.0
python-telegram-bot[job-queue]==20.7
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
